+++ aws-aurora-mysql/variables.tf
@@ -38,7 +38,7 @@ variable "ingress_security_groups" {
 variable "instance_class" {
   type        = string
   description = "See valid instance types [here](https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Managing.Performance.html)"
-  default     = "db.t2.small"
+  default     = "db.t3.medium"
 }
 
 variable "instance_count" {
@@ -164,6 +164,12 @@ variable "iam_database_authentication_enabled" {
   default = false
 }
 
+variable "iam_roles" {
+    type        = list(string)
+    description = "A list of ARNs for the IAM roles to associate to the RDS Cluster."
+    default     = []
+}
+
 variable "db_deletion_protection" {
   type    = string
   default = false
@@ -172,7 +178,7 @@ variable "db_deletion_protection" {
 variable "engine_version" {
   type        = string
   description = "The version of the engine to be used for aurora-mysql."
-  default     = "5.7"
+  default     = "8.0"
 }
 
 variable "ca_cert_identifier" {